Disable friendly ID on artist/tracks
This was causing routing problems, mainly:

 - periods in names (fixed by the routing constraint)
 - url helpers not finding routes for a resource with a slash in it's name

The latter is valid, for example: GZA/Genius, which is correctly decoded
when the request is being routed, eg, routed before decoding, so the
/ isn't treated as a literal when matching routes, but is decoded by the
time it hits the controller in params form.

The inverse doesn't happen when building a route from a resource, it
tries to match the route before escaping, which will fail as only the
first part of the name, before the / will be treated as a param for the
resource, leaving the remainder unmatching a route.

This was generating UrlGenerationError while rendering a list of tracks
or artists, which is more trouble than its worth for now.

Will revisit later, but for now ID based routes are functional and
won't break random bits of pages.