More nesting, scrobbles on everything, with cleaner scoping
Simpler than charts as theres not a group/count step, but this model seems to
work a lot better/more simply, eg switching by resource in the controller
and calling the relation method on the resource, rather than chaining
on a common relation object, trying to keep everything in scope.

Also adds some smart display logic in scroble/chart views, to hide columns
that are part of the scope, eg no artist on an artist top tracks view.

Same for simple breadcrumb/context links, just on scrobble view
as a example for now