Refactor nested resource scope logic in to a concern
A very common pattern appeared in controllers where they
may be nested within another resource, and need to scope
the index action to the resource it's nested within.

Extract out that behaviour into a controller concern to
DRY up the controllers.

This continues to set the outer resource as an instance
variable, so the view can make use of it to either
 - provide navigation through breadcrumbs
 - allow tweaking of the UI based on nesting context