Hide chart columns based on resource nesting
When a track chart is nested within an artist resource it doesn't
make sense to show the artist name (as it will be the same for
all tracks.

For scrobbles#index the artist will be redundent when nested in
an artist resource, and the same for track (nested within an
artist itself) and for a users scrobbles.