Split import logic into testable lib class
Take the creation logic out of the rake task as much as possible so
that it can be tested more easily. In particular there is still a
bug with importing an artist/track by name where it exists, but in
a different case, which should be treated as the same.

This testing will give awareness of regressions introduced while
implementing proper case-aware imports, and allow that to be tested

Not sure the naming is quite right, may be better as a
`UserScrobbleImporter`, init'd with a user, and callable with just
scrobbles, as that is the main use case.