Naively fix Artist/Track import case-sensitivity
Artist and Track names should be case-insensitive. If a record
already exists with a different casing, then that should be used
for the newly imported record.

This was causing errors in real exported scrobble data, despite it
all coming from Last.fm where it should be consistent, as Last.fm
had the same case-sensitivity policy.

Note: this only fixes the problem in import, not any other lookups
using the models, eg, controllers. So `artists/radiohead` will not
find an artist if they're in the DB as `Radiohead`.

Fixing import is an improvement that will allow propering
importing, where lookup in routing can be fixed later, without
needing to reimport.